requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
calmjs==3.4.4
certifi==2023.05.07
cffi==1.16.0
chardet==5.1.0
charset_normalizer==3.1.0
cssselect==1.2.0
decorator==5.1.1
dotenv==1.0.0
exceptiongroup==1.1.2
extruct==0.16.0
h11==0.14.0
html_text==0.5.2
html5lib==1.1
idna==3.4
ijson==3.2.2
isodate==0.6.1
jmespath==1.0.1
js2xml==0.5.0
jstyleson==0.0.2
jsonpath_ng==1.5.3
lxml==4.9.2
mf2py==1.1.3
outcome==1.2.0
packaging==23.1
parsel==1.8.1
ply==3.11
pycparser==2.21
pyparsing==3.1.0
pyRdfa==3.5.3
rdflib==7.0.0
rdflib_jsonld==0.5.0
requests==2.31.0
selenium==4.10.0
setuptools==59.6.0
six==1.16.0
sniffio==1.3.0
socks==1.7.1
sortedcontainers==2.4.0
soupsieve==2.4.1
tqdm==4.65.0
trio==0.22.1
trio_websocket==0.10.3
urllib3==2.0.3
validators==0.20.0
w3lib==2.1.1
webdriver_manager==3.8.6
webencodings==0.5.1
wsproto==1.2.0
